% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_variables.R
\name{abb_to_state}
\alias{abb_to_state}
\title{Convert state abbreviations to names}
\usage{
abb_to_state(.data, state_col)
}
\arguments{
\item{.data}{Dataframe.}

\item{state_col}{String. Name of state column.}
}
\value{
Updated dataframe where all state abbreviations in `state_col` have been
converted to names.
}
\description{
`abb_to_state` changes all state abbreviations in selected
column to state names.
}
\seealso{
[state_to_abb]
}
