% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_col.R
\name{col_to_state}
\alias{col_to_state}
\title{Format state names in column}
\usage{
col_to_state(.data, state_col, abb = TRUE)
}
\arguments{
\item{.data}{Dataframe.}

\item{state_col}{String. Name of state column.}

\item{abb}{Boolean. If `TRUE`, uses abbreviations. If `FALSE`, uses full
state names. Default `TRUE`.}
}
\value{
* If abb is `TRUE`, converts all state names in column to abbreviations
* If abb is `FALSE`, converts all state abbreviations in column to names
}
\description{
Formats all state names in column as either abbreviations or
 full names.
}
\seealso{
[col_to_numeric], [call_to_date], [rename_all_var]
}
