% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_columns.R
\name{col_to_time}
\alias{col_to_time}
\title{Convert datetime column to time}
\usage{
col_to_time(.data, time_col)
}
\arguments{
\item{.data}{Dataframe.}

\item{time_col}{String. Name of time column.}
}
\value{
Converts column to HH:MM string and returns dataframe
}
\description{
`col_to_time` converts a datetime column to a string column containing the
time.
}
\seealso{
[col_to_numeric, col_to_date]
}
