% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colnames_results}
\alias{colnames_results}
\title{Lookup table for result data column names}
\format{
A dataframe with x rows and 8 columns:
 \describe{
   \item{WQX}{WQX}
   \item{MassWateR}{MassWateR}
   \item{WQdashboard}{WQdashboard}
   \item{RI_WW}{Rhode Island Watershed Watch}
   \item{RI_DEM}{Rhode Island DEM}
   \item{MA_BRC}{Blackstone River Coalition}
   \item{ME_DEP}{Maine DEP}
   \item{ME_FOCB}{Friends of Casco Bay}
 }
}
\usage{
colnames_results
}
\description{
Dataframe lookup table containing column names for various water
quality data formats. Each row represents an equivalent column name and each
column represents a different format. Additional formatting notes:
* Column order is specified by the format ##|| at the beginning of each
column name (eg 01||col1, 02||col2).
* If a data format is known to use alternate column names for the same data,
then the alternate names are listed after the preferred/most common name with
the delimeter | (eg colA|colB|colC).
}
\keyword{datasets}
