% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_units.R
\name{convert_unit}
\alias{convert_unit}
\title{Convert unit}
\usage{
convert_unit(x, old_unit, new_unit, unit_format = "wqx")
}
\arguments{
\item{x}{Numeric. Value to convert.}

\item{old_unit}{String. Current unit.}

\item{new_unit}{String. New unit.}

\item{unit_format}{String. Format used for units. Not case sensitive.
Accepted values: WQX, MassWateR, wqdashboard, RI_WW, RI_DEM, MA_BRC, ME_DEP,
ME_FOCB. Default wqx.}
}
\value{
If `old_unit` and `new_unit` are compatible, returns updated value.
If they are incompatible, returns `-999999`.
}
\description{
`convert_unit()` calculates the value for `x` when converted
to a new unit.
}
