% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_var_names.R
\name{fetch_var}
\alias{fetch_var}
\title{Fetch list of variable name substitutions}
\usage{
fetch_var(in_table, in_format, out_format)
}
\arguments{
\item{in_table}{Dataframe.}

\item{in_format, out_format}{String. Column names for the input format
(`in_format`) and output format (`out_format`).}
}
\value{
List containing three items: `old_names`, `new_names`, and
`keep_var`
* `old_names` and `new_names` are paired lists. They contain name
substitutions derived from the `in_format` and `out_format` columns. Matching
pairs between `in_format` and `out_format` are removed.
* `keep_var` is a list of unique variables in the `out_format` column.
Matching pairs between `in_format` and `out_format` are kept.
}
\description{
`fetch_var()` generates list of name substitutions. Helper function for
[format_results()] and [format_sites()].
}
\section{How table data is interpreted}{

* Each column is sorted alphabetically. If variables need to be listed in a
specific order, then append ##|| (eg "01||") to the start of the variable.
Numbers must be exactly two digits. If variable contains "||", then the first
four letters for the variable will be dropped after alphabetizing but
before further processing.
* Alternate variable names can be listed in a cell with the delimiter "|".
Alternate variable names for `in_format` will be included in the output.
Alternate variable names for `out_format` will be dropped.
}

