% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_var_names.R
\name{find_var_names}
\alias{find_var_names}
\title{Find variable names}
\usage{
find_var_names(df, in_format, out_format)
}
\arguments{
\item{df}{Dataframe.}

\item{in_format}{String. Name of column containing input variable names.}

\item{out_format}{String. Name of column containing output variable names.}
}
\value{
List. `old_names` and `new_names` contain paired lists of variable
  names derived from `in_format` and `out_format` columns. Duplicate values
  are not included in lists. `keep_var` contains list of all variables
  listed in `in_format` column.
}
\description{
Generates list of name substitutions from dataframe. Each column
   lists the variables for a given format, and each row lists equivalent
   variables. (UNCLEAR WORDING, REWRITE)

   Alternate variables can be listed in a cell with the delimiter "|".
   Alternate variables in `in_format` will be included in output. Alternate
   variables in `out_format` will only be included if `multiple_out_var` is TRUE.
}
