% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_results.R
\name{format_mwr_results}
\alias{format_mwr_results}
\title{Check and improve formatting for MassWateR result data}
\usage{
format_mwr_results(.data)
}
\arguments{
\item{.data}{Input dataframe.}
}
\value{
Updated dataframe.
}
\description{
`format_mwr_results()` reviews MassWateR result data and checks
for common formatting errors.
* Checks for missing columns and places columns in the correct order
* Updates "Result Value" and "Result Measure Qualifier" for over-detects and
under-detects
* Transfers duplicate values to "QC Reference Value"
}
