% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_results.R
\name{format_results}
\alias{format_results}
\title{Format result data}
\usage{
format_results(
  df,
  in_format,
  out_format,
  date_format = "m/d/Y",
  tz = "America/New_York",
  drop_extra_col = TRUE
)
}
\arguments{
\item{df}{Input dataframe.}

\item{in_format}{String. Name of input format. (word better)}

\item{out_format}{String. Name of desired output format. (word better)}

\item{date_format}{String. Date format, uses lubridate. (word better) Default
value "m/d/Y".}

\item{tz}{Timezone. Default value "America/New_York".}

\item{drop_extra_col}{Boolean. If TRUE, removes any columns that can't be
converted to `out_format`. Default value TRUE.}
}
\value{
Updated dataframe.
}
\description{
Converts water quality result data between formats. (List formats)
}
