% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_results.R
\name{format_wqd_results}
\alias{format_wqd_results}
\title{Check and improve formatting for wqdashboard result data}
\usage{
format_wqd_results(.data, date_format)
}
\arguments{
\item{.data}{Input dataframe}

\item{date_format}{String. Date format. Uses the same formatting as
[lubridate::parse_date_time()]. Default value "m/d/Y".}
}
\value{
Updated dataframe
}
\description{
`format_wqd_results()` reviews wqdashboard result data and
checks for common formatting errors.
* Checks for missing columns
* Checks for unknown parameters, units, qualifiers, and activity type
* Formats Date column as date
}
