% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_format_sites.R
\name{prep_MA_BRC_sites}
\alias{prep_MA_BRC_sites}
\title{Sites from Massachusetts Blackstone River Coalition}
\usage{
prep_MA_BRC_sites(df)
}
\arguments{
\item{df}{Dataframe.}
}
\value{
Updated dataframe.
}
\description{
Helper function for `format_sites` that pre-formats site data
 from the Blackstone River Coalition (MA_BRC).
   * Adds column "STATE" with each site's state
   * Adds column "WATER_DEPTH_M" and converts values in "WATER_DEPTH_FT" to
     meters
}
