% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_MassWateR.R
\name{prep_MassWateR}
\alias{prep_MassWateR}
\title{Prep MassWateR}
\usage{
prep_MassWateR(dat)
}
\arguments{
\item{dat}{Dataframe.}
}
\value{
Updated dataframe.
}
\description{
Converts MassWateR data to standard format. Helper function for
  `format_results`.
 * Transfers "BDL" and "AQL" values from `Result Value` to
   `Result Measure Qualifier`.
 * Data in `QC Reference Value` is copied to its own row.
}
