% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_format_results.R
\name{prep_MassWateR_results}
\alias{prep_MassWateR_results}
\title{Prepare MassWateR Results}
\usage{
prep_MassWateR_results(df)
}
\arguments{
\item{df}{Dataframe.}
}
\value{
Updated dataframe.
}
\description{
Helper function for `format_results`. Standardizes result data
   from MassWateR.
 * Transfers "BDL" and "AQL" values from `Result Value` to
   `Result Measure Qualifier`.
 * Data in `QC Reference Value` is copied to its own row.
}
