% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_ME_FOCB.R
\name{prep_ME_FOCB}
\alias{prep_ME_FOCB}
\title{Prep ME_FOCB Results}
\usage{
prep_ME_FOCB(df, date_format = "m/d/y")
}
\arguments{
\item{df}{Dataframe.}
}
\value{
Updated dataframe.
}
\description{
Pre-formats result data for Maine Friends of Casco Bay (ME_FOCB).
  Adds "Sample Depth Unit" column and pivots table from wide to long.
}
