% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_format_results.R
\name{prep_ME_FOCB_results}
\alias{prep_ME_FOCB_results}
\title{Prep Friends of Casco Bay Results}
\usage{
prep_ME_FOCB_results(df, date_format = "m/d/y")
}
\arguments{
\item{df}{Dataframe.}

\item{date_format}{String. Date format, uses lubridate. (word better)}
}
\value{
Updated dataframe.
}
\description{
Helper function for `format_results`. Standardizes result data
   from the Friends of Casco Bay (ME_FOCB).

  * Adds "Sample Depth Unit" column
  * Pivots table from wide to long.
}
