% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_col.R
\name{rename_col}
\alias{rename_col}
\title{Rename columns}
\usage{
rename_col(.data, old_colnames, new_colnames)
}
\arguments{
\item{.data}{Dataframe}

\item{old_colnames, new_colnames}{List. Columns listed in `old_colnames` will
be renamed to match paired name in `new_colnames`. `old_colnames` and
`new_colnames` must be the same length and in the same order.}
}
\value{
Dataframe with updated column names.
}
\description{
`rename_col()` renames columns by substituting values from `new_colnames` for
values from `old_colnames`.
}
\seealso{
[concat_columns]
}
