% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_var.R
\name{rename_var}
\alias{rename_var}
\title{Rename variable}
\usage{
rename_var(in_var, old_varname, new_varname, multiple = FALSE)
}
\arguments{
\item{in_var}{String. Variable to update.}

\item{old_varname, new_varname}{List. List old variable names in `old_varname`
and new variable names in `new_varname`. `old_varname` and `new_varname` are
a paired list, therefor they must be the same length and in the same order.}

\item{multiple}{Boolean. If `TRUE`, will return all matches for
`in_var`. If `FALSE`, only returns the first match. Default `FALSE`.}
}
\value{
String. If substitution found, provides updated variable name.
Otherwise returns input variable.
}
\description{
`rename_var()` uses a paired list (`old_varname`, `new_varname`) to generate
a list of name substitutions and update the input variable. If no match
found, leaves input variable as-is. Helper function for [rename_all_var()].
}
\seealso{
[rename_all_var]
}
