% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_format_results.R
\name{results_to_MA_BRC}
\alias{results_to_MA_BRC}
\title{Results to Massachusetts Blackstone River Coalition}
\usage{
results_to_MA_BRC(df)
}
\arguments{
\item{df}{Input dataframe.}
}
\value{
Updated dataframe.
}
\description{
Helper function for `format_results` that formats result data for
   the Blackstone River Coalition (MA_BRC).
   * Uses DATE, TIME columns to fill DATE_TIME column
   * Renames variables in PARAMETER column
   * Removes DATE, TIME, and SAMPLE_TYPE columns
}
