% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_format_results.R
\name{results_to_MassWateR}
\alias{results_to_MassWateR}
\title{Results to MassWateR}
\usage{
results_to_MassWateR(df, in_format)
}
\arguments{
\item{df}{Dataframe.}

\item{in_format}{String. Input format.}
}
\value{
Updated dataframe.
}
\description{
Helper function for `format_results`. Formats result data for
   MassWateR.
 * Sets "Result Value" to "BDL" or "AQL" as needed
 * Sets "Result Measure Qualifier" to Q or NA
 * Transfers duplicate values to "QC Reference Value"
}
