% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_units.R
\name{set_units}
\alias{set_units}
\title{Set units}
\usage{
set_units(.data, value, unit, target_unit, unit_format = "wqx", silent = FALSE)
}
\arguments{
\item{.data}{Dataframe}

\item{value}{String. Column with result values.}

\item{unit}{String. Column with units.}

\item{target_unit}{String. Unit to convert `value` and `unit` columns to.}

\item{unit_format}{String. Format used for units. Not case sensitive.
Accepted values: WQX, MassWateR, wqdashboard, RI_WW, RI_DEM, MA_BRC, ME_DEP,
ME_FOCB. Default wqx.}

\item{silent}{Boolean. If `TRUE`, returns a warning message if unable to
convert units. If `FALSE`, does not return a warning.}
}
\value{
Updated dataframe with updated unit and result values. Rows that
can not be converted to `target_unit` are left as-is and in informative
warning or error message is provided.
}
\description{
`set_units()` converts all units and values in target columns to
the selected unit.
}
