% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_format_sites.R
\name{sites_from_MA_BRC}
\alias{sites_from_MA_BRC}
\title{Sites from Massachusetts Blackstone River Coalition}
\usage{
sites_from_MA_BRC(df)
}
\arguments{
\item{df}{Input dataframe.}
}
\value{
Updated dataframe.
}
\description{
Helper function for `format_sites` that formats site data from
   the Blackstone River Coalition (MA_BRC) format.
   * Adds column "STATE" with each site's state
   * Adds column "WATER_DEPTH_M" and converts values in "WATER_DEPTH_FT" to
     meters
}
