% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_format_sites.R
\name{sites_to_MA_BRC}
\alias{sites_to_MA_BRC}
\title{Sites to Massachusetts Blackstone River Coalition}
\usage{
sites_to_MA_BRC(df)
}
\arguments{
\item{df}{Input dataframe.}
}
\value{
Updated dataframe.
}
\description{
Helper function for `format_sites` that formats site data for
   the Blackstone River Coalition (MA_BRC).
   * Converts values in column "WATER_DEPTH_M" to feet and fills empty
     values in column "WATER_DEPTH_FT"
   * Removes columns "STATE", "WATER_DEPTH_M"
}
