% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_units.R
\name{standardize_units}
\alias{standardize_units}
\title{Standardize units}
\usage{
standardize_units(
  .data,
  group,
  value,
  unit,
  unit_format = "wqx",
  warn_only = TRUE
)
}
\arguments{
\item{.data}{Dataframe}

\item{group}{String. Column used to group data.}

\item{value}{String. Column with result values.}

\item{unit}{String. Column with units.}

\item{unit_format}{String. Format used for units. Not case sensitive.
Accepted values: WQX, MassWateR, wqdashboard, RI_WW, RI_DEM, MA_BRC, ME_DEP,
ME_FOCB. Default wqx.}

\item{warn_only}{Boolean. If `TRUE`, returns a warning message if unable to
convert units. If `FALSE`, returns an error message.}
}
\value{
Updated dataframe with updated unit and result values. Rows that
can not be converted to a standard unit are left as-is and in informative
warning or error message is provided.
}
\description{
`standardize_units()` standardizes the unit used for each
parameter in `group`.
}
