% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_units.R
\name{standardize_units_across}
\alias{standardize_units_across}
\title{Standardize units across each row}
\usage{
standardize_units_across(
  .data,
  target_unit,
  unit,
  value,
  unit_format = "wqx",
  warn_only = TRUE
)
}
\arguments{
\item{.data}{Dataframe}

\item{target_unit}{String. Column that contains the units that all other
columns should match.}

\item{unit}{String. Column with units.}

\item{value}{List or string. Column(s) with result values. If multiple
columns are listed, then data in row will only update if the entire row can
be updated.}

\item{unit_format}{String. Format used for units. Not case sensitive.
Accepted values: WQX, MassWateR, wqdashboard, RI_WW, RI_DEM, MA_BRC, ME_DEP,
ME_FOCB. Default wqx.}

\item{warn_only}{Boolean. If `TRUE`, returns a warning message if unable to
convert units. If `FALSE`, returns an error message.}
}
\value{
Updated dataframe. Where possible, units will be standardized across
the row. If units can not be standardized in a row, the units and values for
that row will be left as-is.
}
\description{
`standardize_units_across()` standardizes the units used across
a row of data.
}
