% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_variables.R
\name{state_to_abb}
\alias{state_to_abb}
\title{Convert state names to abbreviations}
\usage{
state_to_abb(.data, state_col)
}
\arguments{
\item{.data}{Dataframe.}

\item{state_col}{String. Name of state column.}
}
\value{
Updated dataframe where all state names in `state_col` have been converted
to abbreviations.
}
\description{
`state_to_abb` changes all state names in selected column to
abbreviations.
}
\seealso{
[abb_to_state]
}
