% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_MassWateR.R
\name{to_MassWateR}
\alias{to_MassWateR}
\title{To MassWateR}
\usage{
to_MassWateR(dat, in_format)
}
\arguments{
\item{dat}{Dataframe.}
}
\value{
Updated dataframe.
}
\description{
Helper function for `format_results` that formats data for
 `MassWateR`.
 * Sets `Result Value` to "BDL" or "AQL" as needed
 * Sets `Result Measure Qualifier` to Q or NA
 * Transfers duplicate values to `QC Reference Value`
}
