% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_variables.R
\name{update_unit}
\alias{update_unit}
\title{Rename units}
\usage{
update_unit(.data, col_name, in_format, out_format)
}
\arguments{
\item{.data}{Dataframe}

\item{col_name}{String. Column name.}

\item{in_format, out_format}{String. Desired input and output formats.
Not case sensitive. Accepted formats:

* WQX

* MassWateR

* wqdashboard

* RI_WW (Rhode Island Watershed Watch)

* RI_DEM (Rhode Island DEM)

* MA_BRC (Blackstone River Coalition)

* ME_DEP (Maine DEP)

* ME_FOCB (Friends of Casco Bay)}
}
\value{
Dataframe with updated units in target column
}
\description{
`update_unit()` converts units in target column to new format.
}
\seealso{
[update_var], [update_unit]
}
