% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{varnames_parameters}
\alias{varnames_parameters}
\title{Lookup table for parameter names}
\format{
A dataframe with x rows and 6 columns:
 \describe{
   \item{WQX}{WQX}
   \item{MassWateR}{MassWateR}
   \item{RI_DEM}{Rhode Island DEM}
   \item{MA_BRC}{Blackstone River Coalition}
   \item{ME_DEP}{Maine DEP}
   \item{ME_FOCB}{Friends of Casco Bay}
 }
}
\usage{
varnames_parameters
}
\description{
Dataframe lookup table containing variable names for parameter type in
various water quality formats. Each row represents an equivalent variable
name and each column represents a different format. Additional formatting
notes:
* If a data format is known to allow multiple variable names for the same
data, then the alternate names are listed after the preferred/most common
name with the delimeter | (eg varA|varB|varC).
}
\keyword{datasets}
