% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_var.R
\name{warn_invalid_var}
\alias{warn_invalid_var}
\title{Warn if Invalid Variables}
\usage{
warn_invalid_var(df, col_name, varlist)
}
\arguments{
\item{df}{Input dataframe.}

\item{col_name}{String. Column name.}

\item{varlist}{List. List of acceptable variable names.}
}
\value{
Updated dataframe.
}
\description{
Checks if column contains variables that aren't included in the
 list of acceptable names.
}
