% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_col.R
\name{col_to_date}
\alias{col_to_date}
\title{Convert column to date or datetime format}
\usage{
col_to_date(.data, date_col, date_format = "m/d/Y", tz = Sys.timezone())
}
\arguments{
\item{.data}{Dataframe.}

\item{date_col}{String. Name of date column.}

\item{date_format}{String. Date format. Uses the same formatting as
[lubridate::parse_date_time()]. Default value "m/d/Y".}

\item{tz}{String. Specifies the timezone used to parse dates. Uses system
timezone as default.}
}
\value{
Converts column to date or datetime and returns dataframe
}
\description{
`col_to_date` converts column to date or datetime format.
* Uses date format by default
* Will use datetime format if time variables are included in `date_format`
}
\seealso{
[col_to_numeric], [col_to_state], [rename_all_var]
}
