% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_to_x.R
\name{col_to_date}
\alias{col_to_date}
\title{Column to Date}
\usage{
col_to_date(df, date_col, date_format = "m/d/Y", tz = Sys.timezone())
}
\arguments{
\item{df}{Input dataframe.}

\item{date_col}{String. Name of date column.}

\item{date_format}{String. Date format. Uses the same formatting as
`lubridate::parse_date_time`. Default value "m/d/Y".}

\item{tz}{String. Specifies the timezone used to parse dates. Defaults to
system timezone.}
}
\description{
Checks if column is formatted as a date. If column is not
  formatted as date, converts column to date type.
}
