% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_to_x.R
\name{col_to_numeric}
\alias{col_to_numeric}
\title{Column to Numeric}
\usage{
col_to_numeric(dat, field)
}
\arguments{
\item{dat}{Dataframe.}

\item{field}{String. Name of column to convert to numeric.}
}
\value{
Updated dataframe.
}
\description{
Checks if column can be converted to numeric format. If TRUE,
 updates column type.
}
