% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_col.R
\name{col_to_numeric}
\alias{col_to_numeric}
\title{Column to Numeric}
\usage{
col_to_numeric(.data, col_name)
}
\arguments{
\item{.data}{Dataframe.}

\item{col_name}{String. Column name.}
}
\value{
Updated dataframe.
}
\description{
Checks if column can be converted to numeric. If all
   entries can be converted with no loss of data, converts column to numeric.
}
