% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_col.R
\name{col_to_numeric}
\alias{col_to_numeric}
\title{Convert column to numeric format}
\usage{
col_to_numeric(.data, col_name)
}
\arguments{
\item{.data}{Dataframe.}

\item{col_name}{String. Column name.}
}
\value{
If all values are numeric, converts column to numeric and returns dataframe.
If there are any non-numeric values, leaves column as-is and returns
dataframe.
}
\description{
If all values in column are numbers, `col_to_numeric()` converts column type
to numeric.
}
\seealso{
[col_to_date], [col_to_state], [rename_all_var]
}
