% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_to_x.R
\name{col_to_state}
\alias{col_to_state}
\title{Format State Columns}
\usage{
col_to_state(df, state_col, full_name = FALSE)
}
\arguments{
\item{df}{Dataframe.}

\item{state_col}{String. Name of state column.}

\item{full_name}{Boolean. If TRUE, updates `state_col` to include full state
names. If FALSE, updates `state_col` to include state abbreviations. Default
FALSE.}
}
\value{
Updated dataframe.
}
\description{
Updates all state names in column to either abbreviations or
 full names.
}
