% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_format_col.R
\name{col_to_state}
\alias{col_to_state}
\title{Format State Columns}
\usage{
col_to_state(.data, state_col, full_name = FALSE)
}
\arguments{
\item{.data}{Dataframe.}

\item{state_col}{String. Name of state column.}

\item{full_name}{Boolean. If \code{TRUE}, formats `state_col` as full
state names. If \code{FALSE}, formats `state_col` as state abbreviations.
Default \code{FALSE}.}
}
\value{
Updated dataframe.
}
\description{
Formats all state names in column as either abbreviations or
 full names.
}
