% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_col.R
\name{concat_columns}
\alias{concat_columns}
\title{Concatenate Columns}
\usage{
concat_columns(df, in_fields, out_field)
}
\arguments{
\item{df}{Dataframe.}

\item{in_fields}{List. Dataframe columns to combine.}

\item{out_field}{String. Name of column to transfer values to.}
}
\value{
Updated dataframe.
}
\description{
Combines values from multiple columns and saves the first
 non-NA value to the output column.
}
