% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_results.R
\name{format_mwr_results}
\alias{format_mwr_results}
\title{Check and improve formatting for MassWateR result data}
\usage{
format_mwr_results(.data, date_format = "m/d/Y")
}
\arguments{
\item{.data}{Input dataframe}

\item{date_format}{String. Date format. Uses the same formatting as
[lubridate::parse_date_time()]. Default value "m/d/Y".}
}
\value{
Updated dataframe
}
\description{
`format_mwr_results()` reviews MassWateR result data and checks
for common formatting errors.
* Checks for missing columns and places columns in the correct order
* Updates "Result Value" and "Result Measure Qualifier" for over-detects and
under-detects
* Transfers duplicate values to "QC Reference Value"
}
