% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_results.R
\name{format_results}
\alias{format_results}
\title{Format result data}
\usage{
format_results(
  df,
  in_format,
  out_format,
  date_format = "m/d/Y",
  tz = Sys.timezone(),
  drop_extra_col = TRUE
)
}
\arguments{
\item{df}{Input dataframe.}

\item{in_format, out_format}{String. Desired input and output formats.
Possible inputs:
* WQX
* MassWateR
* WQdashboard
* RI_WW (Rhode Island Watershed Watch)
* RI_DEM (Rhode Island DEM)
* MA_BRC (Blackstone River Coalition)
* ME_DEP (Maine DEP)
* ME_FOCB (Friends of Casco Bay)}

\item{date_format}{String. Date format. Uses the same formatting as
[lubridate::parse_date_time()]. Default value "m/d/Y".}

\item{tz}{String. Specifies the timezone used to parse dates. Uses system
timezone as default.}

\item{drop_extra_col}{Boolean. If `TRUE`, removes any columns that can't be
converted to `out_format`. Default `TRUE`.}
}
\value{
Updated dataframe.
}
\description{
Converts water quality result data between different formats.
}
