% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_sites.R
\name{format_sites}
\alias{format_sites}
\title{Format site data}
\usage{
format_sites(df, in_format, out_format, drop_extra_col = TRUE)
}
\arguments{
\item{df}{Input dataframe.}

\item{in_format, out_format}{String. Name of input & output formats}

\item{drop_extra_col}{Boolean. If `TRUE`, removes any columns that can't be
converted to `out_format`. Default `TRUE`.}
}
\value{
Updated dataframe.
}
\description{
Converts water quality site data between formats. (List formats)
}
