% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_var.R
\name{rename_all_var}
\alias{rename_all_var}
\title{Rename all variables in column}
\usage{
rename_all_var(df, col_name, old_varname, new_varname)
}
\arguments{
\item{df}{Input dataframe.}

\item{col_name}{String. Column name.}

\item{old_varname}{List. Old variable names; must be same length and in
same order as `new_varname`.}

\item{new_varname}{List. New variable names; must be same length and in
same order as `old_varname`.}
}
\value{
Updated dataframe.
}
\description{
Converts values in column from old variable names to new
  new variable names. If no match found, leaves old variable name.
}
