% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_col.R
\name{rename_col}
\alias{rename_col}
\title{Rename columns}
\usage{
rename_col(df, old_colnames, new_colnames)
}
\arguments{
\item{df}{Dataframe.}

\item{old_colnames}{List. Old column names; must be same length and in
same order as `new_colnames`.}

\item{new_colnames}{List. New column names; must be same length and in
same order as `old_colnames`.}
}
\value{
Updated dataframe.
}
\description{
Rename columns from `old_colnames` to `new_colnames`. If no
  match found, leaves old column name.
}
