% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_columns.R
\name{rename_col}
\alias{rename_col}
\title{Rename columns}
\usage{
rename_col(.data, old_colnames, new_colnames)
}
\arguments{
\item{.data}{Dataframe}

\item{old_colnames, new_colnames}{List. Columns listed in `old_colnames` will
be renamed to match paired name in `new_colnames`. `old_colnames` and
`new_colnames` must be the same length and in the same order.}
}
\value{
Dataframe with updated column names.
}
\description{
`rename_col()` renames dataframe columns by using a paired list to replace
`old_colnames` with `new_colnames`. If no match is found for an existing
column name, it is left as-is.
}
\seealso{
[concat_col]
}
