% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_var.R
\name{rename_var}
\alias{rename_var}
\title{Rename variable}
\usage{
rename_var(in_var, old_varname, new_varname, allow_multiple = FALSE)
}
\arguments{
\item{in_var}{String. Variable to update.}

\item{old_varname}{List. Old variable names; must be same length and in
same order as `new_varname`.}

\item{new_varname}{List. New variable names; must be same length and in
same order as `old_varname`.}

\item{allow_multiple}{Boolean. If \code{TRUE}, will return all matches for
`in_var`. If \code{FALSE}, only returns the first match. Default
\code{FALSE}.}
}
\value{
String. Updated variable name.
}
\description{
Converts value from old variable to new variable name. If no
  match found, leaves old variable name. Helper function for
  `rename_all_var`.
}
