% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_format_results.R
\name{results_from_MA_BRC}
\alias{results_from_MA_BRC}
\title{Results from Massachusetts Blackstone River Coalition}
\usage{
results_from_MA_BRC(df, date_format = "Y-m-d H:M", tz = "America/New_York")
}
\arguments{
\item{df}{Input dataframe.}

\item{date_format}{String. Date format. Uses the same formatting as
`lubridate`. Default value "m/d/Y".}

\item{tz}{String. Timezone. Default value "America/New York".}
}
\value{
Updated dataframe.
}
\description{
Helper function for `format_results` that preformats result
 data from the Blackstone River Coalition (MA_BRC).
   * Adds columns for DATE, TIME, SAMPLE_TYPE
}
