% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_variables.R
\name{warn_invalid_var}
\alias{warn_invalid_var}
\title{Check column for invalid variables}
\usage{
warn_invalid_var(df, col_name, varlist)
}
\arguments{
\item{df}{Dataframe}

\item{col_name}{String. Column name.}

\item{varlist}{List. List of acceptable variable names.}
}
\value{
Updated dataframe.
}
\description{
`warn_invalid_var()` checks if the input column contains any variables that
aren't included in `varlist`.
}
\seealso{
[update_var]
}
